examples/python-uv/src/pyproject.toml
[project]
name = "minimal"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "pendulum>=3.0.0",
]
